% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGGV.R
\name{addpie}
\alias{addpie}
\title{Add allele frequency pie charts to choropleth map}
\usage{
addpie(f, radius = 4)
}
\arguments{
\item{f}{Allele frequency table generated by ggv() (dataframe)}

\item{radius}{Radius for pie charts. (numeric)}
}
\description{
\code{addpie} Adds allele frequency pie charts for all populations in the  
 frequency dataframe.
}
\details{
Adds allele frequency pie charts for all populations in the frequency dataframe.
}
\examples{
freq_table <- ggv()
grid <- seasonAve()
envmap(grid)
addpie(freq_table)
}
