% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RGGV.R
\name{envmap}
\alias{envmap}
\title{Map NCEP-NCAR environmental data}
\usage{
envmap(grid, interior = FALSE, ncol = 18)
}
\arguments{
\item{envvar}{Enviornmental variable name (string)}

\item{season}{Season [winter|summer] (string)}

\item{level}{Data level (string)}

\item{ext}{Function for aggregating data ['min'|'max'|'mean'] (string)}
}
\description{
\code{envmap} Creates a choropleth map displaying seasonal 
environmental data from the NCEP-NCAR database.
}
\details{
Creates a choropleth map displaying seasonal environmental data from the NCEP-NCAR database.
}
\examples{
envmap()
envmap(season="summer")
envmap(envvar='dswrf.surface',season='winter',level='gaussian',ext='max')
}
