% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RGGV.R
\name{ggv}
\alias{ggv}
\title{Map variant in GGV}
\usage{
ggv(rs = NULL, chr = NULL, pos = NULL, db = "hgdp", table = FALSE)
}
\arguments{
\item{rs}{A SNP ID (string)}

\item{chr}{Chromosome (numeric)}

\item{pos}{Physical position (numeric)}

\item{db}{Database "hgdp" (default) or "1kgenomes"}

\item{table}{(boolean)}
}
\description{
\code{ggv} takes either a string \code{rs} or a numeric physical position specified by
arguments \code{chr} and \code{pos}. If neither identifier is specified, a random variant
is mapped. Variants from either the HGDP or 1000 genomes can be retrieved by setting
\code{db = "hgdp"} (default) or \code{db = "1kgenomes"}.
}
\details{
Creates a cluster pie map from variant data in the GGV. Variants can be
retrieved by rs ID or physical position. If neither rs ID or physical
position is specified, a random SNP will be retrieved. The variant data
can be accessed from either the HGDP or 1kgenome databases. To produce a table
instead of a map, set \code{table = TRUE}.
}
\examples{
ggv("rs1834640")
ggv()
ggv(, 6, 130099903, db = "1kgenomes")
}

